requirements.txt

flask 
python-dotenv 
openai
langchain-groq
flask-cors